+++ backendParent/commons/src/main/java/com/backend/commons/service/EmailService.java
@@ -8,7 +8,7 @@ public interface EmailService {
 	void sendEmail(String recipientEmail, String subject, String body, Map<String, File> inlineImages,
 			File attachments) throws Exception;
 
-	void sendOnboardingEmail(String recipientEmail, String fname, String lname, int empId, String password,
+	void sendOnboardingEmail(String recipientEmail, String fname, String lname, Long empId, String password,
 			String origin);
 
 	void sendOtpEmail(String recipientEmail, String otp);
