+++ backendParent/core/src/main/java/com/backend/core/entity/InvoiceTemplate.java
@@ -32,7 +32,7 @@ public class InvoiceTemplate  implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "INVOICEID")
-	private int invoiceId;
+	private Long invoiceId;
 	
 	@Column(name = "ACTIVE")
 	private boolean active;
@@ -47,7 +47,7 @@ public class InvoiceTemplate  implements Serializable {
 		super();
 	}
 
-	public InvoiceTemplate(Tenant tenant, int invoiceId, boolean active, Blob document) {
+	public InvoiceTemplate(Tenant tenant, Long invoiceId, boolean active, Blob document) {
 		super();
 		this.tenant = tenant;
 		this.invoiceId = invoiceId;
@@ -63,11 +63,11 @@ public class InvoiceTemplate  implements Serializable {
 		this.tenant = tenant;
 	}
 
-	public int getInvoiceId() {
+	public Long getInvoiceId() {
 		return invoiceId;
 	}
 
-	public void setInvoiceId(int invoiceId) {
+	public void setInvoiceId(Long invoiceId) {
 		this.invoiceId = invoiceId;
 	}
