+++ backendParent/persistence/src/main/java/com/backend/persistence/dao/CartDao.java
@@ -34,13 +34,13 @@ public class CartDao {
 	@Autowired
 	private ProductService productService;
 	
-	public void insertIntoCart(int productid, int customerid, int quantity) throws Exception {
+	public void insertIntoCart(Long productid, Long customerid, int quantity) throws Exception {
 		try (Connection con = dbUtil.getConnectionInstance()) {
 			PreparedStatement stmt = con
 					.prepareStatement("insert into customercart values(?,?,?,?)");
 			stmt.setString(1, baseService.getTenantInfo().getTenantID());
-			stmt.setInt(2, customerid);
-			stmt.setInt(3, productid);
+			stmt.setLong(2, customerid);
+			stmt.setLong(3, productid);
 			stmt.setInt(4, quantity);
 			stmt.executeUpdate();
 		} catch (Exception ex) {
@@ -49,16 +49,16 @@ public class CartDao {
 		}
 	}
 	
-	public List<CustomerCart> userCartItems(int customerid) throws Exception{
+	public List<CustomerCart> userCartItems(Long customerid) throws Exception{
 		List<CustomerCart> cartItems = new ArrayList<CustomerCart>();
 		try (Connection con = dbUtil.getConnectionInstance()) {
 			PreparedStatement stmt = con.prepareStatement("select productid,quantity from customercart where tenantid=? and customerid=?");
 			stmt.setString(1, baseService.getTenantInfo().getTenantID());
-			stmt.setInt(2, customerid);
+			stmt.setLong(2, customerid);
 			ResultSet rs = stmt.executeQuery();
 			while (rs.next()) {
 				CustomerCart cart = new CustomerCart();
-				cart.setProduct(productService.getProductById(rs.getInt(1)));
+				cart.setProduct(productService.getProductById(rs.getLong(1)));
 				cart.setQuantity(rs.getInt(2));
 				cartItems.add(cart);
 			}
@@ -69,13 +69,13 @@ public class CartDao {
 		return cartItems;
 	}
 	
-	public void removeProductFromCart(int productid, int customerid) throws Exception {
+	public void removeProductFromCart(Long productid, Long customerid) throws Exception {
 		try (Connection con = dbUtil.getConnectionInstance()) {
 			PreparedStatement stmt = con
 					.prepareStatement("delete from customercart where tenantid=? and customerid=? and productid=? ");
 			stmt.setString(1, baseService.getTenantInfo().getTenantID());
-			stmt.setInt(2, customerid);
-			stmt.setInt(3, productid);
+			stmt.setLong(2, customerid);
+			stmt.setLong(3, productid);
 			stmt.executeUpdate();
 		} catch (Exception ex) {
 			logger.error("Exception removing from cart- " + ex);
@@ -83,12 +83,12 @@ public class CartDao {
 		}
 	}
 	
-	public void clearCustomerCart(int customerid) throws Exception {
+	public void clearCustomerCart(Long customerid) throws Exception {
 		try (Connection con = dbUtil.getConnectionInstance()) {
 			PreparedStatement stmt = con
 					.prepareStatement("delete from customercart where tenantid=? and customerid=?");
 			stmt.setString(1, baseService.getTenantInfo().getTenantID());
-			stmt.setInt(2, customerid);
+			stmt.setLong(2, customerid);
 			stmt.executeUpdate();
 		} catch (Exception ex) {
 			logger.error("Exception deleting from cart- " + ex);
@@ -96,14 +96,14 @@ public class CartDao {
 		}
 	}
 	
-	public void updateProductQuantity(int productid, int customerid, int quantity) throws Exception {
+	public void updateProductQuantity(Long productid, Long customerid, int quantity) throws Exception {
 		try (Connection con = dbUtil.getConnectionInstance()) {
 			PreparedStatement stmt = con
 					.prepareStatement("update customercart set quantity=? where tenantid=? and customerid=? and productid=? ");
 			stmt.setInt(1, quantity);
 			stmt.setString(2, baseService.getTenantInfo().getTenantID());
-			stmt.setInt(3, customerid);
-			stmt.setInt(4, productid);
+			stmt.setLong(3, customerid);
+			stmt.setLong(4, productid);
 			stmt.executeUpdate();
 		} catch (Exception ex) {
 			logger.error("Exception inserting into cart- " + ex);
@@ -111,12 +111,12 @@ public class CartDao {
 		}
 	}
 	
-	public int getUserCartCount(int customerid) throws Exception {
+	public int getUserCartCount(Long customerid) throws Exception {
 		try (Connection con = dbUtil.getConnectionInstance()) {
 			PreparedStatement stmt = con
 					.prepareStatement("select count(*) from customercart where tenantid=? and customerid=? ");
 			stmt.setString(1, baseService.getTenantInfo().getTenantID());
-			stmt.setInt(2, customerid);
+			stmt.setLong(2, customerid);
 			ResultSet rs = stmt.executeQuery();
 			while (rs.next()) {
 				return rs.getInt(1);
