+++ backendParent/persistence/src/main/java/com/backend/persistence/entity/Orders.java
@@ -38,10 +38,10 @@ public class Orders implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "ORDERID")
-	private int orderId;
+	private Long orderId;
 	
 	@Column(name = "CUSTOMERID", nullable = false)
-	private int customerId;
+	private Long customerId;
 	
 	@Column(name = "STATUS")
 	private String status;
@@ -59,7 +59,7 @@ public class Orders implements Serializable {
 	private boolean couponapplied;
 	
 	@Column(name = "COUPONID")
-	private int couponId;
+	private Long couponId;
 	
 	@Column(name = "PAYMENTMODEID")
 	private int paymentModeId;
@@ -68,7 +68,7 @@ public class Orders implements Serializable {
 	private long orderPlacedTime;
 	
 	@Column(name = "EMPLOYEEID", nullable = true)
-	private int employeeId;
+	private Long employeeId;
 	
 	@OneToOne(mappedBy = "orderId", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
 	private OrderInvoice invoice;
@@ -77,7 +77,7 @@ public class Orders implements Serializable {
 	private List<OrderDetails> orderDetails;
 	
 	@Column(name = "CUSTOMERADDRESSID")
-	private int customerAddressId;
+	private Long customerAddressId;
 	
 	@Column(name = "DELIVERYCHARGE")
 	private int deliveryCharge;
@@ -90,19 +90,19 @@ public class Orders implements Serializable {
 		this.tenant = tenant;
 	}
 
-	public int getOrderId() {
+	public Long getOrderId() {
 		return orderId;
 	}
 
-	public void setOrderId(int orderId) {
+	public void setOrderId(Long orderId) {
 		this.orderId = orderId;
 	}
 
-	public int getCustomerId() {
+	public Long getCustomerId() {
 		return customerId;
 	}
 
-	public void setCustomerId(int customerId) {
+	public void setCustomerId(Long customerId) {
 		this.customerId = customerId;
 	}
 
@@ -146,19 +146,19 @@ public class Orders implements Serializable {
 		this.couponapplied = couponapplied;
 	}
 
-	public int getCouponId() {
+	public Long getCouponId() {
 		return couponId;
 	}
 
-	public void setCouponId(int couponId) {
+	public void setCouponId(Long couponId) {
 		this.couponId = couponId;
 	}
 
-	public int getEmployeeId() {
+	public Long getEmployeeId() {
 		return employeeId;
 	}
 
-	public void setEmployeeId(int employeeId) {
+	public void setEmployeeId(Long employeeId) {
 		this.employeeId = employeeId;
 	}
 
@@ -182,11 +182,11 @@ public class Orders implements Serializable {
 		this.paymentModeId = paymentModeId;
 	}
 
-	public int getCustomerAddressId() {
+	public Long getCustomerAddressId() {
 		return customerAddressId;
 	}
 
-	public void setCustomerAddressId(int customerAddressId) {
+	public void setCustomerAddressId(Long customerAddressId) {
 		this.customerAddressId = customerAddressId;
 	}
